soccer kick: smooth, deliberate run-up (pose interpolation + slower cadence)
The run-up flailed because the gait was quantized to integer authored ticks and
crammed ~3 steps into a 16-tick sprint played 1:1 (~11 steps/s). Fix it three ways
(not just "slow down"):

- POSE INTERPOLATION: the kicker now samples the authored keyframes at FRACTIONAL
  ticks — SoccerPenaltyKickPose::joints_at (and PenaltyPhysicsKick::joints_at for
  the experimental path) interpolate the 13 joint transforms between the two
  bracketing frames (translation lerp + Quat nlerp). kicker_frame + boxes_at are
  now f32, so motion is continuous BETWEEN keyframes instead of snapping.
- SLOWER RATE: the run-up advances at RUNUP_PLAYBACK_RATE (0.5) authored ticks per
  game tick; the shot script holds the charge ~96 ticks so it still reaches the
  cocked pose.
- GENTLER CADENCE: 2 strides (was 3) + smaller maxima (stride 1.0->0.9, arm
  0.95->0.75, knee 1.1->1.0).

Filmstrip markers re-grounded on the ~2x-longer timeline; kicker/kick-animation
tests updated for the f32 signatures. Module + soccer lib tests green.

+++ apps/axiom-gallery/tests/soccer_penalty_kick_animation.rs
@@ -69,7 +69,7 @@ fn the_motion_spec_compiles_and_has_nine_phases_in_order() {
 fn setup_places_the_kicker_behind_the_ball() {
     // In the rendered WORLD, the setup/idle kicker stands behind the ball: its pelvis
     // sits at a larger z (nearer the camera) than the ball on the penalty spot.
-    let boxes = KickerRig::new().boxes_at(PHASE_SAMPLE_TICKS[SETUP] as u32);
+    let boxes = KickerRig::new().boxes_at(PHASE_SAMPLE_TICKS[SETUP] as f32);
     let pelvis = boxes.iter().find(|b| b.label == "kicker.pelvis").expect("pelvis box");
     assert!(pelvis.center.z > PENALTY_SPOT_Z, "kicker stands behind the ball: pelvis z={}", pelvis.center.z);
 }
